Split signature algorithms in separate files
