+++ src/meson.build
@@ -14,7 +14,10 @@ pkcs11_provider_sources = [
   'provider.c',
   'random.c',
   'session.c',
-  'signature.c',
+  'sig/signature.c',
+  'sig/rsasig.c',
+  'sig/ecdsa.c',
+  'sig/eddsa.c',
   'skeymgmt.c',
   'slot.c',
   'store.c',
